// types/building.ts
export interface BuildingData {
    structuralSystem?: string;
    buildingType?: string;
    buildingHeight?: string;
    constructionArea?: string;
    foundationArea?: string;
    basementFloorNumber?: string;
    totalFloorNumber?: string;
    buildingUsageClass?: string;
    buildingHeightClass?: string;
    buildingImportanceFactor?: string;
}